modified sondes.yaml from JO-v0.11.X and resulting changes
Following changes in the JOANNE dataset led to changes in the sondes.yaml:

(a) ASPEN processing was carried out on the D-files and not the D*_P files. This led to some differences in the soundings and therefore the files output by ASPEN are now *_QC, not *_PQC files, which contain some changes. This causes changes in the Level-1 files, which are the input for the Jo-QC tests, and thus, the qc_flag. The changes in the sonde flags (also for sondes that came into or went out of existence) is because Level-1 files changed. Therefore, input for the QC before Level-2 also changed. This caused the changes. But the same, original sonde_ids have been maintained. One sonde that got added to HALO-0209 was given the last s number of the flight instead of whatever sequence it was coming in (I think 27 or something), and one sonde that got removed from HALO-0215 also took with it the sonde_id that it was originally given. This means that this sonde id is not present in the FPS files (or for that matter in JOANNE) anymore.

(b) Additionally, the way launch_time is attributed to sondes has changed in JOANNE. Initially, launch_time was minimum of the time values for a sounding, but now it is the value of the attribute launch_time for every sounding. For most cases, the two are the same, but not for some, and so launch_time changes can be seen in the sondes.yaml

The above changes in the sondes.yaml result in changes to the flight-phase-segmentation (FPS) files. The changes to the FPS files include:

(a) changes to FLAGs of sondes
(b) addition of TTFS tags in irregularities where the segment start is no longer a minute from first sonde launch_time, because the first sonde launch_time changed in the modified sondes.yaml