sort keys in compiled yaml file in human friendly order
The order in the compiled yaml files doesn't have a semantic meaning,
but it should be kept mostly fixed to simplify the search for changes.
Befor this change the order has been alphabetically, which is not very
friendly to a human reader. This change list a few manually selected
keys in a preferred order and uses this order for sorting. Additional
keys may also be present and are appended in alphabetical order (as
before).