Use nicknames instead of names
The use of RFxx flight names has always been confusing and ambiguous.
The topic has been discussed already during the field campaign and we
concluded that we offically want to refer to flights only by their
flight_id. Still, those RF-names have made it until this place and keep
popping up here and there. This commit get's rid of those in the
official flight listing. However, as it might be fun to have another
name for each flight, we've got (hopefully less ambiguous) nicknames for each.